Revert "bemhtml: experimental cache"
This reverts commit 7a404417dfff99906f669e6f085f58ab3dda5a75.
