--- blocks-common/i-bem/__html/benchmarks/cache-partial-example-benchmark.json
@@ -1,72 +1,0 @@
-{
-    "cache": "key",
-    "block": "b-page",
-    "title": "Pseudo link",
-    "head": [
-        { "elem": "css", "url": "example.css"},
-        { "elem": "css", "url": "example.ie.css", "ie": "lt IE 8" },
-        { "block": "i-jquery", "elem": "core" },
-        { "elem": "js", "url": "example.js" }
-    ],
-    "content": { "link": "content" },
-    "links": {
-      "content": [
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        },
-        {
-            "block": "b-link",
-            "mods": { "pseudo" : "yes", "togcolor" : "yes", "color": "green" },
-            "url": "#",
-            "target": "_blank",
-            "title": "Click me",
-            "content" : "This pseudo link changes its color after click"
-        }
-    ]
-  }
-}
