+++ blocks-common/i-bem/__html/test/files/basic-block.bemhtml
