--- blocks-common/i-bem/__html/test/files/cache/basic.html
@@ -1,1 +1,0 @@
-<div class="hello"><div class="how"></div><p class="are"></p><div class="you"></div></div>
