--- blocks-common/i-bem/__html/test/files/cache/basic.json
@@ -1,19 +1,0 @@
-{
-  "cache": "key",
-  "block": "hello",
-
-  "content": [{
-    "block": "how"
-  }, {
-    "link": "are"
-  }, {
-    "block": "you"
-  }],
-
-  "links": {
-    "are": {
-      "tag": "p",
-      "block": "are"
-    }
-  }
-}
