+++ blocks-common/i-bem/__html/test/files/comments.bemhtml
