+++ blocks-common/i-bem/__html/test/files/empty-attrs.bemhtml
