+++ blocks-common/i-bem/__html/test/files/gh-232.bemhtml
