+++ blocks-common/i-bem/__html/test/files/gh-239.bemhtml
