+++ blocks-common/i-bem/__html/test/files/gh-254.bemhtml
