+++ blocks-common/i-bem/__html/test/files/gh-289.bemhtml
