+++ blocks-common/i-bem/__html/test/files/local-var.html
