+++ blocks-common/i-bem/__html/test/files/one-symbol-elem.json
