+++ blocks-common/i-bem/__html/test/files/single-mix.bemhtml
