+++ blocks-common/i-bem/__html/test/i-bem-test.js
@@ -15,9 +15,9 @@ suite('i-bem block and others', function() {
   function unit(name, file, raw) {
     test(name, function() {
       var contents = {
-        src: readFile('i-bem/' + file + '.bemhtml'),
-        data: JSON.parse(readFile('i-bem/' + file + '.json')),
-        dst: readFile('i-bem/' + file + '.html')
+        src: readFile(file + '.bemhtml'),
+        data: JSON.parse(readFile(file + '.json')),
+        dst: readFile(file + '.html')
       };
 
       assert.equal(
