+++ package.json
@@ -9,7 +9,7 @@
   },
   "devDependencies": {
     "mocha": "1.2.x",
-    "benchmark": "~1.0.0",
+    "benchmark": "~0.3.0",
     "microtime": "~0.3.1"
   },
   "scripts": {
